+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+### 6.1.10
+
+- Add support for GML3.2 to `WMSFeatureInfoResponseType`
+- Add missing properties in the `toJSON` function of the VectorTileLayer: `highlight` and `debug`
+
 ### 6.1.9
 
 - Fixes an issue with WMS FeatureInfo and MapServer
-- Add missing properties in the `toJSON` function of the VectorTileLayer: `highlight` and `debug`
 
 ### 6.1.8